Create an easyweb rmarkdown website

This function takes as input a two column table with a set of `tag`s and `value`s that are used to construct a rmarkdown website that you can then modify yourself and publish. This website will help you get your online presence started!

easyweb(web_tsv, path = tempdir(), publish = FALSE,
  overwrite = FALSE)

Arguments

web_tsv
The path to a tab-separated file with `tag`s on the first column and `value`s on the second. Check the example table in the package for a template.

path
The path where you website files will be saved to.

publish
Whether to publish your website (currently not implemented).

overwrite
Whether to overwrite the files.

Value

The path to your website.

Examples

## Example easyweb table
options(width = 200)
easyweb_example

#>              tag                                                                                                                             value
#> 1           name                                                                                                           Leonardo Collado-Torres
#> 2        twitter                                                                                                                        fellgernon
#> 3         github                                                                                                                       lcolladotor
#> 4       linkedin                                                                                                                          lcollado
#> 5  googlescholar                                                                                                                      h57-MykAAAAJ
#> 6            doi                                                                                                         10.1038/s41593-018-0197-y
#> 7            doi                                                                                                                  10.1038/nbt.3838
#> 8   presentation [CDSB2019 workshop](https://github.com/ComunidadBioInfo/cdsb2019). Taught how to make R packages (TIB2019, Cuernavaca, July 2019)
#> 9   presentation                          LIIGH 2019. recount-brain and BrainSEQ Phase II scientific talk (LIIGH seminar, Juriquilla, August 2019)
#> 10         email                                                                                                             lcolladotor@gmail.com
#> 11         email                                                                                                              leo.collado@libd.org
#> 12       address                                                                                                                Baltimore, MD, USA
#> 13          blog                                                                                                     https://lcolladotor.github.io
#> 14         phone                                                                                                                   +1 123 456 7899
#> 15         phone                                                                                                                   +1 012 345 6789
#> 16      interest                                                                                                                     R programming
#> 17      interest                                                                                                                          Genomics
#> 18  introduction                                   Hello! I'm interested in teaching others how to build R packages, particularly Latin Americans.

## Create an example website directory where the files will be saved
example_website_path <- file.path(tempdir(), 'easyweb')
dir.create(example_website_path, showWarnings = FALSE)
example_tsv <- file.path(example_website_path, 'web.tsv')

## Save the example table as a file
write.table(easyweb_example, file = example_tsv, sep = '\t', quote = FALSE, row.names = FALSE)

## Build the website
easyweb(example_tsv, path = example_website_path, overwrite = TRUE)

#> Warning: EOF within quoted string

#> Writing 2 Bibtex entries ... 

#> OK
#> Results written to file ‘/var/folders/cx/n9s558kx6fb7jf5z_pgszgb80000gn/T//RtmppuGcG0/easyweb/citations.bib’

#>   |                                                                         |                                                                 |   0%  |                                                                         |.................................................................| 100%
#>   ordinary text without R code
#> 
#> 
#> /Applications/RStudio.app/Contents/MacOS/pandoc/pandoc +RTS -K512m -RTS contact.utf8.md --to html4 --from markdown+autolink_bare_uris+ascii_identifiers+tex_math_single_backslash+smart --output contact.html --email-obfuscation none --standalone --section-divs --template /Library/Frameworks/R.framework/Versions/3.6/Resources/library/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --css styles.css --include-before-body /var/folders/cx/n9s558kx6fb7jf5z_pgszgb80000gn/T//RtmppuGcG0/rmarkdown-str2c8936a0bea6.html --variable navbar=1 --variable body_padding=64 --variable header_padding=69 --variable 'theme:paper' --include-in-header /var/folders/cx/n9s558kx6fb7jf5z_pgszgb80000gn/T//RtmppuGcG0/rmarkdown-str2c897344a04e.html --mathjax --variable 'mathjax-url:https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' 
#>   |                                                                         |                                                                 |   0%  |                                                                         |.................................................................| 100%
#>   ordinary text without R code
#> 
#> 
#> /Applications/RStudio.app/Contents/MacOS/pandoc/pandoc +RTS -K512m -RTS index.utf8.md --to html4 --from markdown+autolink_bare_uris+ascii_identifiers+tex_math_single_backslash+smart --output index.html --email-obfuscation none --standalone --section-divs --template /Library/Frameworks/R.framework/Versions/3.6/Resources/library/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --css styles.css --include-before-body /var/folders/cx/n9s558kx6fb7jf5z_pgszgb80000gn/T//RtmppuGcG0/rmarkdown-str2c896130d83d.html --variable navbar=1 --variable body_padding=64 --variable header_padding=69 --variable 'theme:paper' --include-in-header /var/folders/cx/n9s558kx6fb7jf5z_pgszgb80000gn/T//RtmppuGcG0/rmarkdown-str2c894dc4bea4.html --mathjax --variable 'mathjax-url:https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --metadata pagetitle=index.utf8.md 
#>   |                                                                         |                                                                 |   0%  |                                                                         |.................................................................| 100%
#>   ordinary text without R code
#> 
#> 
#> /Applications/RStudio.app/Contents/MacOS/pandoc/pandoc +RTS -K512m -RTS presentations.utf8.md --to html4 --from markdown+autolink_bare_uris+ascii_identifiers+tex_math_single_backslash+smart --output presentations.html --email-obfuscation none --standalone --section-divs --template /Library/Frameworks/R.framework/Versions/3.6/Resources/library/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --css styles.css --include-before-body /var/folders/cx/n9s558kx6fb7jf5z_pgszgb80000gn/T//RtmppuGcG0/rmarkdown-str2c8944ebcbf9.html --variable navbar=1 --variable body_padding=64 --variable header_padding=69 --variable 'theme:paper' --include-in-header /var/folders/cx/n9s558kx6fb7jf5z_pgszgb80000gn/T//RtmppuGcG0/rmarkdown-str2c89549c6bc8.html --mathjax --variable 'mathjax-url:https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --metadata pagetitle=presentations.utf8.md 
#>   |                                                                         |                                                                 |   0%  |                                                                         |......................                                           |  33%
#>   ordinary text without R code
#> 
#>   |                                                                         |...........................................                      |  67%
#> label: publications (with options) 
#> List of 5
#>  $ results: chr "asis"
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ eval   : language file.exists("citations.bib")
#> 
#>   |                                                                         |.................................................................| 100%
#>    inline R code fragments
#> 
#> 
#> /Applications/RStudio.app/Contents/MacOS/pandoc/pandoc +RTS -K512m -RTS publications.utf8.md --to html4 --from markdown+autolink_bare_uris+ascii_identifiers+tex_math_single_backslash+smart --output publications.html --email-obfuscation none --standalone --section-divs --template /Library/Frameworks/R.framework/Versions/3.6/Resources/library/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --css styles.css --include-before-body /var/folders/cx/n9s558kx6fb7jf5z_pgszgb80000gn/T//RtmppuGcG0/rmarkdown-str2c891d981742.html --variable navbar=1 --variable body_padding=64 --variable header_padding=69 --variable 'theme:paper' --include-in-header /var/folders/cx/n9s558kx6fb7jf5z_pgszgb80000gn/T//RtmppuGcG0/rmarkdown-str2c896c1efd3b.html --mathjax --variable 'mathjax-url:https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --metadata pagetitle=publications.utf8.md 

#> 
#> Output created: index.html

#> [1] "/var/folders/cx/n9s558kx6fb7jf5z_pgszgb80000gn/T//RtmppuGcG0/easyweb"